split OpenPMDParticleReader into atomic state and charge state subclass
